용식 Board댓글부분 새파이롤 컴퍼넌트 분리 and 반복문
